requirements.txt
defusedxml==0.4.1
dj-database-url==0.3.0
dj-static==0.0.6
Django==1.8.6
django-bmemcached==0.2.2
django-cors-headers==1.1.0
django-crispy-forms==1.5.2
django-debug-toolbar==1.4
django-extensions==1.5.9
django-filter==0.11.0
django-formtools==1.0
django-toolbelt==0.0.1
djangorestframework==3.3.1
gunicorn==19.3.0
oauthlib==1.0.3
Pillow==3.0.0
psycopg2==2.6.1
PyJWT==1.4.0
python-binary-memcached==0.24.6
python-memcached==1.57
python-openid==2.2.5
python-social-auth==0.2.13
python3-openid==3.0.9
requests==2.8.1
requests-oauthlib==0.5.0
sendgrid==1.4.0
sendgrid-django==1.3.0
six==1.10.0
smtpapi==0.2.0
sqlparse==0.1.18
static3==0.6.1
wheel==0.24.0
